Basic compression and decompression functionality; RLE broken

+++ bwt.h
@@ -3,15 +3,15 @@
 
 #include "util.h"
 
-/* Apply the Burrows-Wheeler transform on the data, writing the result
+/* Apply the Burrows-Wheeler transform to the data, writing the result
  * to the output array. The length of the output array must be length+1.
  */
-void bwt(symbol *data, symbol *output, int length);
+void bwt(symbol *data, symbol *output, size_t length);
 
-/* Apply the inverse Burrows-Wheeler transform on the data, writing the
+/* Apply the inverse Burrows-Wheeler transform to the data, writing the
  * result to the output array. The length of the output array shall be
  * length-1.
  */
-void unbwt(symbol *data, symbol *output, int length);
+void unbwt(symbol *data, symbol *output, size_t length);
 
 #endif
