+++ huffman.h
@@ -20,7 +20,7 @@ typedef struct huf_tree huf_tree;
 
 /* Build a Huffman tree from the given data.
  */
-huf_tree huf_build_tree(symbol* data, int length);
+huf_tree huf_build_tree(symbol* data, size_t length);
 
 /* Write the representation of the given Huffman tree to a file.
  */
@@ -28,7 +28,7 @@ void huf_write_tree(file *f, huf_tree tree);
 
 /* Encode the data using the given Huffman tree and write it to a file.
  */
-void huf_encode(file *f, symbol *data, int length, huf_tree tree);
+void huf_encode(file *f, symbol *data, size_t length, huf_tree tree);
 
 /* Build a Huffman tree from its representation stored in a file.
  */
@@ -38,9 +38,9 @@ huf_tree huf_read_tree(file *f);
  * A pointer to the newly allocated output array is passed to *output and
  * the length of the decoded block is returned.
  */
-int huf_decode(file *f, symbol **output, huf_tree tree);
+size_t huf_decode(file *f, symbol **output, huf_tree tree);
 
-/* Destroy the given huffman tree.
+/* Destroy the given Huffman tree.
  */
 void huf_free_tree(huf_tree tree);
