+++ rle.h
@@ -7,12 +7,12 @@
  * identical symbols is followed by a repeat length. A pointer to the newly
  * allocated output array is passed to *output and its length is returned.
  */
-int rle(symbol *data, symbol **output, int length, int symbols);
+size_t rle(symbol *data, symbol **output, size_t length, size_t symbols);
 
 /* Decode the data encoded using the run-length encoding. A pointer to
  * the newly allocated output array is passed to *output and its length is
  * returned.
  */
-int unrle(symbol *data, symbol **output, int length, int symbols);
+size_t unrle(symbol *data, symbol **output, size_t length, size_t symbols);
 
 #endif
